backend/requirements.txt
# FastAPI和相关依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF处理
PyMuPDF==1.23.8
PyPDF2==3.0.1

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 配置和环境变量
python-dotenv==1.0.0

# 日志
loguru==0.7.2

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 代码质量
black==23.11.0
isort==5.12.0
flake8==6.1.0

# 类型检查
mypy==1.7.1

# 数据库支持
neo4j-driver==5.13.0
# 移除SQLAlchemy依赖，改用Neo4j
# sqlalchemy==2.0.23